+++ tc-default_themes/src/default_quotes.rs
@@ -1,4 +1,4 @@
-use tc_user_config_loader::quote::Quote;
+use tc_user_config_loader::models::quote::Quote;
 
 pub(crate) fn init_default_quotes() -> Vec<Quote> {
     vec![
