+++ tc-user_config_loader/src/clock_types/analog_clock.rs
@@ -1,6 +1,6 @@
 use ratatui::widgets::Paragraph;
 
-use crate::clock::Clock;
+use crate::models::clock::Clock;
 
 pub struct AnalogClock {
     // The static ascii art for the clock face
