+++ tc-user_config_loader/src/clock_types/digital_clock.rs
@@ -1,4 +1,4 @@
-use crate::clock::Clock;
+use crate::models::clock::Clock;
 use ratatui::widgets::Paragraph;
 
 pub struct DigitalClock {
