+++ tc-user_config_loader/src/lib.rs
@@ -1,8 +1,3 @@
-pub mod clock;
-pub mod clock_face_loader;
 pub mod clock_types;
-pub mod colorscheme;
-pub mod colorscheme_loader;
-pub(crate) mod error;
-pub mod quote;
-pub mod quote_loader;
+pub mod loader;
+pub mod models;
