+++ tc-user_config_loader/src/models/quote.rs
